added specific api route for each product

+++ models/Book.php
@@ -6,7 +6,7 @@ require_once('../models/Product.php');
 
 class Book extends Product
 {
-    protected int $weight;
+    protected float $weight;
 
     public function getWeight()
     {
